requirements.txt
# Vedda Translator Requirements
# Core dependencies
flask>=3.1.2
flask-cors>=6.0.1
requests>=2.32.5

# Data management
pandas>=2.3.3
openpyxl>=3.1.5

# Database - MongoDB
pymongo>=4.6.0

# Development
python-dotenv>=1.1.1